Move extension to utility module.
Spec out some initial tests for the utility.
