+++ utility/src/main/java/org/oppia/android/util/extensions/BundleExtensions.kt
@@ -1,4 +1,4 @@
-package org.oppia.android.app.utility
+package org.oppia.android.util.extensions
 
 import android.os.Bundle
 import com.google.protobuf.ByteString
